+++ src/onegov.core/onegov/core/cli/commands.py
@@ -3,20 +3,23 @@ import email
 import mailbox
 import os
 import platform
+import shutil
 import subprocess
 import sys
 
 from cached_property import cached_property
 from mailthon.middleware import TLS, Auth
+from fnmatch import fnmatch
 from onegov.core.cli.core import command_group, pass_group_context
+from onegov.core.cache import lru_cache
 from onegov.core.mail import Postman
+from onegov.core.orm import Base, SessionManager
 from onegov.core.upgrade import get_tasks
 from onegov.core.upgrade import get_upgrade_modules
 from onegov.core.upgrade import RawUpgradeRunner
 from onegov.core.upgrade import UpgradeRunner
 from onegov.server.config import Config
 from smtplib import SMTPRecipientsRefused
-from uuid import uuid4
 
 
 #: onegov.core's own command group
@@ -124,7 +127,10 @@ def sendmail(group_context,
     sys.exit(exit_code)
 
 
-@cli.command()
+@cli.command(context_settings={
+    'matches_required': False,
+    'default_selector': '*'
+})
 @click.argument('server')
 @click.argument('remote-config')
 @click.option('--confirm/--no-confirm', default=True,
@@ -162,154 +168,136 @@ def transfer(group_context,
     print("Parsing the remote application configuration")
 
     remote_dir = os.path.dirname(remote_config)
-    remote_config = Config.from_yaml_string(
-        subprocess.check_output([
-            "ssh", server, "-C", "sudo cat '{}'".format(remote_config)
-        ])
-    )
+
+    try:
+        remote_config = Config.from_yaml_string(
+            subprocess.check_output([
+                "ssh", server, "-C", "sudo cat '{}'".format(remote_config)
+            ])
+        )
+    except subprocess.CalledProcessError:
+        sys.exit(1)
 
     remote_applications = {a.namespace: a for a in remote_config.applications}
 
-    # storages may be shared between applications, so we need to keep track
-    fetched = set()
+    # some calls to the storage transfer may be repeated as applications
+    # share folders in certain configurations
+    @lru_cache(maxsize=None)
+    def transfer_storage(remote, local, glob='*'):
+        send = f"ssh {server} -C 'sudo nice -n 10 tar cz {remote}/{glob}'"
+        send = f"{send} --absolute-names"
+        recv = f"tar xz  --strip-components {remote.count('/') + 1} -C {local}"
 
-    def download_folder(remote, local):
-        tar_filename = '/tmp/{}.tar.gz'.format(uuid4().hex)
+        if shutil.which('pv'):
+            recv = f'pv --name "{remote}/{glob}" -r -b | {recv}'
 
-        subprocess.check_output([
-            'ssh', server, '-C', "sudo tar czvf '{}' -C '{}' .".format(
-                tar_filename, remote
-            )
-        ])
+        subprocess.check_output(f'{send} | {recv}', shell=True)
 
-        subprocess.check_output([
-            'scp',
-            '{}:{}'.format(server, tar_filename),
-            '{}/transfer.tar.gz'.format(local)
-        ])
+    def transfer_database(remote_db, local_db, schema_glob='*'):
+        send = f"ssh {server} sudo -u postgres nice -n 10 pg_dump {remote_db}"
+        send = f"{send} --no-owner --no-privileges --clean --if-exists"
+        send = f"{send} --quote-all-identifiers --no-sync"
 
-        subprocess.check_output([
-            'tar', 'xzvf', '{}/transfer.tar.gz'.format(local),
-            '-C', local
-        ])
+        recv = f"psql -d {local_db} -v ON_ERROR_STOP=1"
 
-        subprocess.check_output([
-            'ssh', server, '-C', "sudo rm '{}'".format(tar_filename)
-        ])
+        if schema_glob != '*':
+            query = 'SELECT schema_name FROM information_schema.schemata'
 
-        subprocess.check_output([
-            'rm', '{}/transfer.tar.gz'.format(local)
-        ])
+            lst = f'sudo -u postgres psql {remote_db} -t -c "{query}"'
+            lst = f"ssh {server} '{lst}'"
 
-    # filter out namespaces on demand
-    for appcfg in group_context.appcfgs:
+            schemas = subprocess.check_output(lst, shell=True)
+            schemas = (s.strip() for s in schemas.decode('utf-8').splitlines())
+            schemas = (s for s in schemas if s)
+            schemas = (s for s in schemas if fnmatch(s, schema_glob))
 
-        if appcfg.configuration.get('disable_transfer'):
-            print("Skipping {}, transfer disabled".format(appcfg.namespace))
+            send = f'{send} --schema {" --schema ".join(schemas)}'
+
+        if platform.system() == 'Linux':
+            recv = f"sudo -u postgres {recv}"
+
+        if shutil.which('pv'):
+            recv = f'pv --name "{remote_db}@postgres" -r -b | {recv}'
+
+        subprocess.check_output(f'{send} | {recv}', shell=True)
+
+    def transfer_storage_of_app(local_cfg, remote_cfg):
+
+        remote_storage = remote_cfg.configuration.get('filestorage')
+        local_storage = local_cfg.configuration.get('filestorage')
+
+        if remote_storage.endswith('OSFS') and local_storage.endswith('OSFS'):
+            local_fs = local_cfg.configuration['filestorage_options']
+            remote_fs = remote_cfg.configuration['filestorage_options']
+
+            remote_storage = os.path.join(remote_dir, remote_fs['root_path'])
+            local_storage = os.path.join('.', local_fs['root_path'])
+
+            transfer_storage(
+                remote_storage, local_storage, glob=f'global-*')
+
+            transfer_storage(
+                remote_storage, local_storage, glob=f'{local_cfg.namespace}*')
+
+    def transfer_depot_storage_of_app(local_cfg, remote_cfg):
+
+        depot_local_storage = 'depot.io.local.LocalFileStorage'
+        remote_backend = remote_cfg.configuration.get('depot_backend')
+        local_backend = local_cfg.configuration.get('depot_backend')
+
+        if local_backend == remote_backend == depot_local_storage:
+            local_depot = local_cfg.configuration['depot_storage_path']
+            remote_depot = remote_cfg.configuration['depot_storage_path']
+
+            remote_storage = os.path.join(remote_dir, remote_depot)
+            local_storage = os.path.join('.', local_depot)
+
+            transfer_storage(
+                remote_storage, local_storage, glob=f'{local_cfg.namespace}*')
+
+    def transfer_database_of_app(local_cfg, remote_cfg):
+        if 'dsn' not in remote_cfg.configuration:
+            return
+
+        if 'dsn' not in local_cfg.configuration:
+            return
+
+        # on an empty database we need to create the extensions first
+        mgr = SessionManager(local_cfg.configuration['dsn'], Base)
+        mgr.create_required_extensions()
+
+        local_db = local_cfg.configuration['dsn'].split('/')[-1]
+        remote_db = remote_cfg.configuration['dsn'].split('/')[-1]
+
+        transfer_database(
+            remote_db, local_db, schema_glob=f'{local_cfg.namespace}*')
+
+    # transfer the data
+    for local_cfg in group_context.appcfgs:
+
+        if local_cfg.namespace not in remote_applications:
             continue
 
-        if appcfg.namespace not in remote_applications:
-            print("Skipping {}, not found on remote".format(appcfg.namespace))
+        if local_cfg.configuration.get('disable_transfer'):
+            print(f"Skipping {local_cfg.namespace}, transfer disabled")
             continue
 
-        remotecfg = remote_applications[appcfg.namespace]
+        remote_cfg = remote_applications[local_cfg.namespace]
+
+        print(f"Fetching {remote_cfg.namespace}")
 
         if not no_filestorage:
-            if remotecfg.configuration.get('filestorage').endswith('OSFS'):
-                assert appcfg.configuration.get('filestorage').endswith('OSFS')
+            transfer_storage_of_app(local_cfg, remote_cfg)
+            transfer_depot_storage_of_app(local_cfg, remote_cfg)
 
-                print("Fetching remote filestorage")
+        if not no_database:
+            transfer_database_of_app(local_cfg, remote_cfg)
 
-                local_fs = appcfg.configuration['filestorage_options']
-                remote_fs = remotecfg.configuration['filestorage_options']
-
-                remote_storage = os.path.join(
-                    remote_dir, remote_fs['root_path'])
-                local_storage = os.path.join('.', local_fs['root_path'])
-
-                if ':'.join((remote_storage, local_storage)) in fetched:
-                    continue
-
-                download_folder(remote_storage, local_storage)
-
-                fetched.add(':'.join((remote_storage, local_storage)))
-
-            remote_backend = remotecfg.configuration.get('depot_backend')
-            local_backend = appcfg.configuration.get('depot_backend')
-
-            if remote_backend == 'depot.io.local.LocalFileStorage':
-                assert local_backend == remote_backend
-
-                print("Fetching depot storage")
-
-                local_depot = appcfg.configuration['depot_storage_path']
-                remote_depot = remotecfg.configuration['depot_storage_path']
-
-                remote_storage = os.path.join(remote_dir, remote_depot)
-                local_storage = os.path.join('.', local_depot)
-
-                if ':'.join((remote_storage, local_storage)) in fetched:
-                    continue
-
-                download_folder(remote_storage, local_storage)
-
-                fetched.add(':'.join((remote_storage, local_storage)))
-
-        if 'dsn' in remotecfg.configuration:
-            assert 'dsn' in appcfg.configuration
-
-            print("Fetching remote database")
-
-            local_db = appcfg.configuration['dsn'].split('/')[-1]
-            remote_db = remotecfg.configuration['dsn'].split('/')[-1]
-            database_dump = subprocess.check_output([
-                'ssh', server, '-C',
-                (
-                    "sudo -u postgres pg_dump "
-                    "--no-owner --no-privileges --clean {}".format(
-                        remote_db
-                    )
-                ),
-            ])
-
-            with open('dump.sql', 'w') as f:
-                # add cascade to all drop schema lines
-                # http://www.postgresql.org/message-id/50EC9574.9060500
-                for line in database_dump.splitlines():
-                    line = line.decode('utf-8')
-
-                    if line.startswith('DROP SCHEMA'):
-                        if 'CASCADE' not in line:
-                            if 'DROP SCHEMA extensions' not in line:
-                                line = line.replace(';', ' CASCADE;')
-
-                    f.write(line + os.linesep)
-
-            try:
-                if platform.system() == 'Darwin':
-                    subprocess.check_output([
-                        "cat dump.sql | psql {}".format(local_db)
-                    ], stderr=subprocess.STDOUT, shell=True)
-                elif platform.system() == 'Linux':
-                    subprocess.check_output([
-                        (
-                            "cat dump.sql "
-                            "| sudo -u postgres psql -d {}".format(
-                                local_db
-                            )
-                        )
-                    ], stderr=subprocess.STDOUT, shell=True)
-                else:
-                    raise NotImplementedError
-            except subprocess.CalledProcessError as e:
-                lines = e.output.decode('utf-8').rstrip('\n').split('\n')
-
-                if lines[-1].startswith('WARNING: errors ignored on restore:'):
-                    pass
-                else:
-                    raise
-            finally:
-                subprocess.check_output(['rm', 'dump.sql'])
+    if not shutil.which('pv'):
+        print("")
+        print("TIP: To get better output with transfer rates, install pv:")
+        print("brew install pv")
+        print("")
 
 
 @cli.command(context_settings={'default_selector': '*'})